Issue #47 - resolve conflict with pull/48.
