+++ app/pillar/handler/tag.go
@@ -2,39 +2,40 @@ package handler
 
 import (
 	"encoding/json"
-	"github.com/coralproject/pillar/pkg/crud"
+	"github.com/coralproject/pillar/pkg/model"
+	"github.com/coralproject/pillar/pkg/service"
 	"net/http"
 )
 
 func GetTags(w http.ResponseWriter, r *http.Request) {
-	jsonObject := crud.Tag{}
+	jsonObject := model.Tag{}
 	json.NewDecoder(r.Body).Decode(&jsonObject)
 
 	// Write content-type, status code and payload
 	w.Header().Set("Content-Type", "application/json")
-	dbObject, err := crud.GetTags()
+	dbObject, err := service.GetTags()
 	doRespond(w, dbObject, err)
 }
 
 func UpsertTag(w http.ResponseWriter, r *http.Request) {
 	//Get the tag from request
-	jsonObject := crud.Tag{}
+	jsonObject := model.Tag{}
 	json.NewDecoder(r.Body).Decode(&jsonObject)
 
 	// Write content-type, status code and payload
 	w.Header().Set("Content-Type", "application/json")
-	dbObject, err := crud.UpsertTag(&jsonObject)
+	dbObject, err := service.UpsertTag(&jsonObject)
 	doRespond(w, dbObject, err)
 }
 
 func DeleteTag(w http.ResponseWriter, r *http.Request) {
 	//Get the tag from request
-	jsonObject := crud.Tag{}
+	jsonObject := model.Tag{}
 	json.NewDecoder(r.Body).Decode(&jsonObject)
 
 	// Write content-type, status code and payload
 	w.Header().Set("Content-Type", "application/json")
-	err := crud.DeleteTag(&jsonObject)
+	err := service.DeleteTag(&jsonObject)
 	doRespond(w, nil, err)
 }
