+++ pkg/service/fixtures/assets.json
